backend/requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
Flask-SQLAlchemy
mysql-connector-python
bcrypt
python-dotenv